# backend/docs/contracts/identity.openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fatura Ótica Identity — contrato Sprint 01",
    "version": "0.1.0",
    "description": "DESIGN-FIRST: endpoints ainda não implementados. Empresa configurada por ambiente, URL única. Sem tenant em payload de login. Contrato deve convergir ao OpenAPI exportado da API antes de client TS. Segurança da aplicação inclui vínculo atual e filial além de bearer."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Exemplo local; porta via configuração"
    }
  ],
  "paths": {
...
    "/api/v1/identity/users": {
      "post": {
        "operationId": "identityCreatePendingProfile",
        "tags": [
          "Identity"
        ],
        "description": "S1-04A1: bearer JWT validado; tenant/autor confiáveis do ambiente/principal. Owner ou BranchManager ativo e atualmente autorizado à filial existente/ativa cria perfil pending. branchId é proveniência, nunca concessão; sem senha, token, papéis ou acesso operacional.",
        "responses": {
          "201": {
            "description": "Perfil criado; status sempre pending; email normalizado; sem credenciais ou concessões.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "identity.validation_failed: campos inválidos ou propriedades não declaradas.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "identity.invalid_credentials: bearer ausente/inválido ou principal/tenant incompatível.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "identity.access_denied: conta/capacidade negada ou filial inexistente, inativa, não autorizada ou de outro tenant; resposta uniforme.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "identity.email_already_exists: email normalizado duplicado no próprio tenant; nenhuma criação parcial.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Falha inesperada; Problem genérico, sem detalhes internos; transação revertida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fatura_refresh",
        "description": "Opaco HttpOnly; host-only; SameSite=Lax; Secure em HTTPS produção. Nunca retorna no JSON."
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Token anti-CSRF associado à sessão; validado junto ao Origin allowlist."
      }
    },
    "schemas": {
...
      "CreateProfileRequest": {
        "type": "object",
        "description": "Cadastro pendente fechado: nenhum campo de tenant, autor, status, papel ou credencial é aceito. Tenant/autor vêm do contexto autenticado confiável.",
        "additionalProperties": false,
        "required": [
          "email",
          "name",
          "branchId"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "branchId": {
            "type": "string",
            "description": "Filial de contexto/destino, existente, ativa e autorizada. Preservada como CreatedForBranchId; não concede acesso operacional. UUID não vazio.",
            "format": "uuid"
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "email",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "blocked"
            ]
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "traceId",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "minLength": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
}